requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
pykrx
pandas
openpyxl
google-generativeai
python-multipart
